requirements.txt
requests
datetime
beautifulsoup4
flask
psycopg2-binary
